+++ docs/identifier.rst
@@ -1,0 +1,5 @@
+==========
+Identifier
+==========
+
+.. automodule:: pyked.identifier
