+++ docs/index.rst
@@ -45,7 +45,7 @@ Code API
    chemked
    converters
    validation
-   orcid
+   identifier
