+++ src/pyked/identifier.py
@@ -1,10 +1,41 @@
-"""Validation helpers for chemical species identifiers."""
+"""Validation helpers for chemical species and author identifiers.
 
+Provides InChI/SMILES parsing checks backed by RDKit and ORCID lookups
+against the ORCID public API.
+"""
+
+import httpx2 as httpx
 from rdkit import Chem, rdBase
 from rdkit.Chem import inchi
 
 INCHI_PREFIX = "InChI="
 
+headers = {"Accept": "application/json"}
+
+def search_orcid(orcid):
+    """
+    Search the ORCID public API
+
+    Specfically, return a dictionary with the personal details
+    (name, etc.) of the person associated with the given ORCID
+
+    Args:
+        orcid (`str`): The ORCID to be searched
+
+    Returns:
+        `dict`: Dictionary with the JSON response from the API
+
+    Raises:
+        `~httpx.HTTPStatusError`: If the given ORCID cannot be found, an
+            `~httpx.HTTPStatusError` is raised with status code 404
+    """
+    url = f"https://pub.orcid.org/v3.0/{orcid}/person"
+    r = httpx.get(url, headers=headers)
+    r.raise_for_status()
+    return r.json()
+
+
+"""Validation helpers for chemical species identifiers."""
 
 def normalize_inchi(value: str):
     """Return an InChI with the conventional ``InChI=`` prefix.
